requirements.txt
pywebview>=5.0
zai-sdk>=0.2.2
openai>=1.30
python-docx>=1.1.0
pyinstaller>=6.0